Rename add liquidity examples
