+++ src/utils/constants.ts
@@ -336,3 +336,4 @@ export const DEFAULT_FUND_MANAGMENT = {
 export const DEFAULT_USERDATA = '0x';
 
 export const API_ENDPOINT = 'https://api-v3.balancer.fi';
+export const TEST_API_ENDPOINT = 'https://test-api-v3.balancer.fi';
